fsck.gfs2: Change bitmap_type variables to int
This patch changes all references of "q" from a uint8_t to a normal
int. This will allow us to check for bad return codes at some point.

+++ gfs2/fsck/fsck.h
@@ -53,7 +53,7 @@ struct dir_info
 struct dir_status {
 	uint8_t dotdir:1;
 	uint8_t dotdotdir:1;
-	uint8_t q;
+	int q;
 	uint32_t entry_count;
 };
